refactor(nlp): move stubParse to @dxos/nlp/testing and use LLM parseText in POS story
stubParse is a demo/test fallback, not a production tagger, so it now lives
only behind the testing entrypoint. PosTranscription story tags via parseText
over the edge AI service (same AiServiceTestingPreset('edge-remote') pattern
as the crawler story), so it now requires edge credentials.
