+++ packages/core/compute/nlp/src/index.ts
@@ -5,4 +5,4 @@
 export * from './Document';
 export * from './align';
 export * from './hash';
-export { type Parser, parseText, stubParse } from './parse';
+export { type Parser, parseText } from './parse';
