+++ packages/core/compute/nlp/src/testing/index.ts
@@ -4,4 +4,4 @@
 
 // Import directly from `stub` (not `parse`) so the testing entrypoint stays isolated to the
 // offline tagger and does not pull the live-parser/AI stack into story and test bundles.
-export { stubParse } from '../stub';
+export { stubParse } from './stub';
