+++ packages/core/compute/nlp/src/testing/stub.test.ts
